requirements.txt
aiohttp
Cython
IPython
librosa
matplotlib
nbformat
numpy
Pillow
pygame
pyserial
scipy
setuptools
torch
torchaudio
adafruit-circuitpython-neopixel
sounddevice